+++ CHANGELOG.md
@@ -1,10 +1,20 @@
+## 1.6.0
+
+Added option to set `requiresWiFi` on the `BackgroundDownloadTask`, which ensures the task won't
+start downloading unless a WiFi network is available. By default `requiresWiFi` is false, and 
+downloads will use the cellular (or metered) network if WiFi is not available, which may incur
+cost.
+
 ## 1.5.0
 
-Added `allTasks` method to get a list of running tasks. Use `allTaskIds` to get a list of taskIds only.
+Added `allTasks` method to get a list of running tasks. Use `allTaskIds` to get a list of taskIds
+only.
 
 ## 1.4.2
 
-Added note to README referring to an issue (and [fix](https://github.com/firebase/flutterfire/issues/9689#issuecomment-1304491789)) where the firebase plugin interferes with the downloader 
+Added note to README referring to an issue (
+and [fix](https://github.com/firebase/flutterfire/issues/9689#issuecomment-1304491789)) where the
+firebase plugin interferes with the downloader
 
 ## 1.4.1
 
@@ -20,16 +30,18 @@ Added option to use an event listener instead of (or in addition to) callbacks
 
 ## 1.2.0
 
-Added FileDownloader.download as a convenience method for simple downloads. This method's Future completes only after the download has completed or failed, and can be used for simple downloads where status and progress checking is not required.
+Added FileDownloader.download as a convenience method for simple downloads. This method's Future
+completes only after the download has completed or failed, and can be used for simple downloads
+where status and progress checking is not required.
 
 ## 1.1.0
 
-Added headers and metaData fields to the BackgroundDownloadTask. Headers will be added to the request, and metaData is ignored but may be helpful to the user
+Added headers and metaData fields to the BackgroundDownloadTask. Headers will be added to the
+request, and metaData is ignored but may be helpful to the user
 
 ## 1.0.2
 
 Replaced Ktor client with a basic Kotlin implementation
-
 
 ## 1.0.0